Repairs and adds Dashboard
Adds an the beginning of an experimental cubic profile. I may not continue pursuing this because I need to devote a lot of effort toward making trapezoidal perfect and working on the prototype drivebase.

Trapezoidal profiling works, however it does not leave the mode after completion of the profile.

Max Velocity = 13.5
Max Accel = 8
Kv =0.08
Ka = 0.05

I'm not sure Ka is actually doing anything.

Action Items:

Make it possible to generate profile from different initial velocities and final velocities
Make it possible to change distance on the fly
Integrate PID